# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Arabic Academy API - Malaysian-curriculum-inspired Arabic learning
 * OpenAPI spec version: 0.1.0
 */
export type CurriculumInfoId = typeof CurriculumInfoId[keyof typeof CurriculumInfoId];


export const CurriculumInfoId = {
  malaysia: 'malaysia',
  uae: 'uae',
} as const;

/**
 * Text localized in the six UI languages
 */
export interface LocalizedText {
  ar: string;
  en: string;
  ms: string;
  id: string;
  fr: string;
  es: string;
}

export interface CurriculumInfo {
  id: CurriculumInfoId;
  name: LocalizedText;
  description: LocalizedText;
  flag: string;
  gradeCount: number;
}

export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  error: string;
}

export interface Overview {
  totalGrades: number;
  totalUnits: number;
  totalLessons: number;
  totalWords: number;
  totalExercises: number;
  skills: string[];
}

export interface LessonSummary {
  id: string;
  unitId: string;
  gradeId: string;
  order: number;
  title: LocalizedText;
  /** Which of the four skills this lesson trains (listening, speaking, reading, writing) */
  skills: string[];
  exerciseCount: number;
  wordCount: number;
}

export interface Unit {
  id: string;
  gradeId: string;
  order: number;
  title: LocalizedText;
  /** Theme keyword (e.g. letters, greetings, food, animals) */
  theme: string;
  lessons: LessonSummary[];
}

export interface Grade {
  id: string;
  number: number;
  title: LocalizedText;
  description: LocalizedText;
  /** Accent hex color for the grade */
  color: string;
  /** Icon keyword (e.g. letters, family, kitchen) */
  icon: string;
  unitCount: number;
  lessonCount: number;
  units: Unit[];
}

export interface VocabItem {
  id: string;
  /** Fully vocalized Arabic word or phrase */
  arabic: string;
  transliteration: string;
  translation: LocalizedText;
  /** Emoji used as visual cue in data only (UI may render differently) */
  emoji: string;
}

export interface DialogueLine {
  speaker: string;
  arabic: string;
  translation: LocalizedText;
}

export interface ExercisePair {
  /** Arabic side */
  a: string;
  /** Match target (translation or related Arabic) */
  b: string;
}

export interface Exercise {
  id: string;
  /** One of mcq, listening, speaking, writing, matching, arrange */
  type: string;
  /** One of listening, speaking, reading, writing */
  skill: string;
  prompt: LocalizedText;
  /**
     * Arabic text to listen to, read aloud, or copy
     * @nullable
     */
  arabicText?: string | null;
  options?: string[];
  /** @nullable */
  correctIndex?: number | null;
  pairs?: ExercisePair[];
  /**
     * Expected answer for writing/arrange exercises
     * @nullable
     */
  answer?: string | null;
  /**
     * Optional image URL shown above writing/arrange exercises
     * @nullable
     */
  imageUrl?: string | null;
}

export interface Lesson {
  id: string;
  unitId: string;
  gradeId: string;
  order: number;
  title: LocalizedText;
  skills: string[];
  intro: LocalizedText;
  vocabulary: VocabItem[];
  dialogue: DialogueLine[];
  exercises: Exercise[];
}

export interface ChatMessage {
  /** user or assistant */
  role: string;
  content: string;
}

export interface ChatRequest {
  messages: ChatMessage[];
  /** UI language code (ar, en, ms, id, fr, es) */
  language: string;
  /**
     * Optional learner grade for level-appropriate replies
     * @nullable
     */
  gradeNumber?: number | null;
}

export interface ChatReply {
  reply: string;
}

export type ListGradesParams = {
/**
 * Filter grades by curriculum ID
 */
curriculum?: ListGradesCurriculum;
};

export type ListGradesCurriculum = typeof ListGradesCurriculum[keyof typeof ListGradesCurriculum];


export const ListGradesCurriculum = {
  malaysia: 'malaysia',
  uae: 'uae',
} as const;